'Noa Physics Module' Fix
From the commit 518b018 over in Minefork, I mistakenly skipped over the Mojang-specific changes to the physics module in Noa. I discovered this while trying to walk through mushrooms in the latest build. Tracked it down to that commit, realized I didn't completely check the diff of the source, like I am thoroughly now. I was just eyeballing it, and that's when you miss things! It's annoying to have to map the whole module out by hand, but you end up not missing any (hopefully) details that may be vital to the implementation of the game.

I feel like if Mojang wanted to, they could port Java Edition line-by-line to another platform, but (logically on their behalf) it probably isn't worthwhile for them to put that much work into doing that, it would be a heck ton of work for not as many returns, compared to say adding a new feature with the current stacks instead.

Kind of a tangent, the concept is that it is worth it to dig deep when what you want is something that comes from that, say like reverse-engineering a closed-source and abandoned converter or something like that :)

I have hopes that I will be able to make a mod into Noa that allows you to export the block data as plain NBT, which would allow for worlds going between versions. I've mentioned this before, but we really are getting closer to that now, than ever before!

Once this entire build setup/stack is up and running properly, I'm going to look into reverting my own custom modernizations too, so then you can start from the ground up with making a modded version of the Classic Minecraft app. I'll have to go back through the earlier commits before the module decompilations, then work those changes back into the now individually-defined modules that we have now (much easier to manage, not too much to address!)

I want to make some niceity changes to the code as well, but maybe that can be a new fork on top of the vanilla code too. Same with the TypeScript and or ESM parts, should I do that before or after enabling easy modding support? Questions to figure out over time :) I do want to modernize all of the dependencies too, I'd want to do that gracefully though.

So yeah, maybe only add modding support once everything is type-safe with TS and or JSDoc, and with upgraded dependencies that are type-safe too.