Full Game Inputs Diff
Ok, this is how I'm going to handle these edits for now. It makes it challenging because that would be a separate npm package on it's own. I may have to do that too, making a Mojang-specific for of that, like noa, but we'll see. It would be helpful to have all of the edits in a single spot.