Components Return Types
These are temporarily intentionally omitting `any`-typed fields, as I want those to show as type errors until they are properly typed. `any` is 'evil'! ...when not necessary. If `any` is intentional, then that's okay. But most of the cases where `any` currently is, those aren't as intentional as they could potentially be.