ent-comp Main Definition File
The tsconfig couldn't find the definition file, so this addresses that. In the package itself, it is `ECS.js`, but that doesn't seem to map to definition file resolution. If it's in the package itself it might be able to, but usually you'd have to reference it in the package.json also. At least I usually do.