ent-comp Private Field Typings Removal
Removing the typings that are actually private in the scope of the original implementation code, yet I had to change it's visibility for when making the typings themselves.

The typings still could use a bit of help. The generics and or Component usage might not be totally correct. It works at the compiler level at the main part of the file, for the most part, so it's at least fairly close. Better than `any` at least, and for now.